Tidied classes and allowed for full object passing
Clients can now pass objects - network packets - to the server, and the server can take action to be implemented by the user. This includes being able to send the packet to the destination specified in the packet.

Also tided overall structure of classes including where streams where being opened and closed - there may be an issue down the road with this as the streams are ot being closed but this needs to be investigated - for now the demostration classes shows how the project can be used to create a very basic texting application.